backend/requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
channels==4.0.0
daphne==4.0.0
